api/requirements/requirements.txt
amqp==2.3.2
ansible==2.6.5
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
decorator==4.3.0
Django==2.1.2
django-celery-beat==1.1.1
django-filter==2.0.0
django-redis-cache==1.7.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-bulk==0.2.1
drf-nested-routers==0.90.2
drf-yasg==1.10.2
flex==6.13.2
future==0.16.0
gitdb2==2.0.4
GitPython==2.1.11
idna==2.7
inflection==0.3.1
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
jsonpointer==2.0
kombu==4.2.1
markdown2==2.3.6
MarkupSafe==1.0
openapi-codec==1.3.2
paramiko==2.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.3.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
PyNaCl==1.3.0
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
rfc3987==1.3.8
ruamel.yaml==0.15.71
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
smmap2==2.0.4
strict-rfc3339==0.7
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
validate-email==1.3
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
channels==2.1.5
channels_redis==2.3.1
gunicorn==19.9.0
djangorestframework-jwt==1.11.0
mysqlclient==1.4.2
